# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Crowdfunding Platform API
  version: 0.3.0
  description: API cho nền tảng gây quỹ cộng đồng Góp Mầm.
servers:
  - url: http://localhost:4000/api/v1
    description: Backend local
paths:
...
  /campaigns/{id}/milestones:
    get:
      operationId: listCampaignMilestones
      summary: Danh sách mốc tiến độ của một chiến dịch
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Danh sách mốc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneListResult"
    post:
      operationId: createMilestone
      summary: Tạo mốc tiến độ mới
      description: >-
        Chỉ chủ dự án hoặc admin; chặn khi chiến dịch đã kết thúc
        (`success`/`failed`/`cancelled`/`ended`). Ghi audit `milestone.create`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneRequest"
      responses:
        "201":
          description: Đã tạo mốc
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Chiến dịch đã kết thúc, không thể sửa mốc
  /campaigns/{id}/milestone-revisions:
    get:
      operationId: listMilestoneRevisions
      summary: Lịch sử thay đổi kế hoạch sau khi phát hành (công khai, 50 bản mới nhất)
      description: Không trả người sửa; chỉ giá trị cũ/mới của trường thay đổi và lý do.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Danh sách thay đổi
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    milestoneId: { type: string, format: uuid }
                    milestoneTitle: { type: string }
                    reason: { type: string }
                    oldValues: { $ref: "#/components/schemas/MilestoneRevisionValues" }
                    newValues: { $ref: "#/components/schemas/MilestoneRevisionValues" }
                    createdAt: { type: string, format: date-time }
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/updates:
    get:
      operationId: listCampaignUpdates
      summary: Nhật ký tiến độ công khai của chiến dịch (mới nhất trước), kèm chứng từ chi tiêu
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - name: offset
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: Danh sách bài cập nhật
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      allOf:
                        - $ref: "#/components/schemas/MilestoneUpdate"
                        - type: object
                          properties:
                            milestoneTitle: { type: string }
                  total: { type: integer }
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/progress:
    get:
      operationId: getCampaignProgress
      summary: Tổng hợp tiến độ chiến dịch (tính từ giao dịch đã hoàn tất)
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Tổng hợp tiến độ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignProgress"
        "404":
          $ref: "#/components/responses/NotFound"
  /milestones/{id}:
    patch:
      operationId: updateMilestone
      summary: Cập nhật một mốc tiến độ
      description: >-
        Ghi audit `milestone.update`. Khi chiến dịch đã phát hành
        (`approved/active/paused/success/ended`), đổi tên/mô tả/hạn/ngân sách bắt buộc
        có `changeReason` (≥ 10 ký tự); thay đổi được lưu vào `milestone_revisions`
        (công khai qua `GET /campaigns/{id}/milestone-revisions`). Đổi hạn/ngân sách thì
        thông báo `milestone_rescheduled` cho người ủng hộ/theo dõi và đặt lại bộ nhắc quá hạn.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMilestoneRequest"
      responses:
        "200":
          description: Đã cập nhật
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
        "400":
          description: Chiến dịch đã phát hành mà thiếu `changeReason`
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Chiến dịch đã kết thúc, hoặc đổi hạn/ngân sách của mốc đã hoàn thành
    delete:
      operationId: removeMilestone
      summary: Xóa một mốc tiến độ (chỉ trước khi phát hành)
      description: Ghi audit `milestone.delete`. Sau khi phát hành không xóa được — đổi hạn/ngân sách kèm lý do.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Đã xóa
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Chiến dịch đã phát hành hoặc đã kết thúc
  /milestones/{id}/complete:
    post:
      operationId: completeMilestone
      summary: Đánh dấu mốc đã hoàn thành
      description: Ghi audit `milestone.complete`.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Đã đánh dấu hoàn thành
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Milestone"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /milestones/{id}/updates:
    get:
      operationId: listMilestoneUpdates
      summary: Danh sách bài cập nhật/chứng từ chi tiêu của một mốc
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Danh sách bài cập nhật
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MilestoneUpdate"
    post:
      operationId: addMilestoneUpdate
      summary: Thêm bài cập nhật/chứng từ chi tiêu cho một mốc
      description: >-
        Chỉ chủ dự án/admin, khi chiến dịch ở `approved/active/paused/success/ended`.
        Có `expenseAmount > 0` thì bắt buộc kèm `receipts` (ảnh `expense_receipt` do
        chính người đăng tải lên, mỗi ảnh chỉ dùng một lần) và tổng chi đã báo cáo
        không vượt số tiền đã huy động được xác nhận. Ghi audit `milestone_update.create`,
        thông báo cho người ủng hộ/theo dõi.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMilestoneUpdateRequest"
      responses:
        "201":
          description: Đã thêm bài cập nhật
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneUpdate"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          description: Thiếu chứng từ khi khai chi tiêu, chứng từ không hợp lệ hoặc trùng
        "409":
          description: >-
            Chiến dịch chưa phát hành/đã thất bại/bị hủy, tổng chi vượt số đã huy động,
            hoặc chứng từ đã dùng cho báo cáo khác
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Forbidden:
      description: Không có quyền thực hiện hành động này
    NotFound:
      description: Không tìm thấy tài nguyên
...
  schemas:
...
    Milestone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campaignId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
          nullable: true
        targetDate:
          type: string
          format: date-time
          nullable: true
        budget:
          type: number
          nullable: true
        sortOrder:
          type: integer
        isCompleted:
          type: boolean
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateMilestoneRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        targetDate:
          type: string
          format: date-time
        budget:
          type: number
          minimum: 0
        sortOrder:
          type: integer
          minimum: 0
          maximum: 10000
    MilestoneRevisionValues:
      type: object
      description: Chỉ có các trường đã thay đổi
      properties:
        title: { type: string }
        description: { type: string, nullable: true }
        targetDate: { type: string, format: date-time, nullable: true }
        budget: { type: number, nullable: true }
    UpdateMilestoneRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
        targetDate:
          type: string
          format: date-time
        budget:
          type: number
          minimum: 0
        sortOrder:
          type: integer
          minimum: 0
          maximum: 10000
        changeReason:
          type: string
          minLength: 10
          maxLength: 500
          description: Bắt buộc khi chiến dịch đã phát hành; công khai trong lịch sử thay đổi
    MilestoneListResult:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Milestone"
        total:
          type: integer
    MilestoneUpdate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        milestoneId:
          type: string
          format: uuid
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        expenseAmount:
          type: number
          nullable: true
        attachments:
          type: array
          description: Chứng từ chi tiêu (ảnh hóa đơn/biên lai)
          items:
            $ref: "#/components/schemas/MilestoneUpdateAttachment"
        createdAt:
          type: string
          format: date-time
    MilestoneUpdateAttachment:
      type: object
      properties:
        id: { type: string, format: uuid }
        mediaFileId: { type: string, format: uuid }
        url: { type: string, example: "/api/v1/media/receipts/3f1c….jpg" }
        caption: { type: string, nullable: true }
        sortOrder: { type: integer }
    CreateMilestoneUpdateRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 10
          maxLength: 5000
        imageUrl:
          type: string
          description: Ảnh hiện trường (link http(s) hoặc ảnh đã tải lên)
        expenseAmount:
          type: number
          minimum: 0
        receipts:
          type: array
          maxItems: 6
          description: Bắt buộc khi `expenseAmount > 0`
          items:
            type: object
            required: [mediaId]
            properties:
              mediaId: { type: string, format: uuid, description: "`id` từ `POST /media/images` với `purpose=expense_receipt`" }
              caption: { type: string, maxLength: 200 }
    CampaignProgress:
      type: object
      required:
        [totalMilestones, completedMilestones, totalBudget, totalExpense, totalRaised, backerCount]
      properties:
        totalMilestones:
          type: integer
        completedMilestones:
          type: integer
        totalBudget:
          type: number
        totalExpense:
          type: number
        totalRaised:
          type: number
          description: Tổng giao dịch `completed`, không tính từ counter người dùng nhập
        backerCount:
          type: integer
        overdueMilestones:
          type: integer
          description: Số mốc quá hạn chưa hoàn thành (kể cả đã giải trình)
        lastUpdateAt:
          type: string
          format: date-time
          nullable: true
        revisionCount:
          type: integer
          description: Số lần sửa kế hoạch sau khi phát hành
        transparency:
          type: string
          enum: [no_plan, on_track, late, late_explained]
        milestoneStates:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              state:
                type: string
                enum: [completed, on_track, overdue, overdue_explained]
                description: >-
                  `overdue`: quá hạn chưa hoàn thành; `overdue_explained`: đã có bài cập nhật
                  cho mốc sau hạn
...